tcp: separating implementation WIP
