+++ api/net/tcp/connection_states.hpp
@@ -15,14 +15,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#pragma once
 #ifndef NET_TCP_CONNECTION_STATES_HPP
 #define NET_TCP_CONNECTION_STATES_HPP
 
-#include <net/tcp.hpp>
+#include <net/tcp/tcp.hpp>
+
+namespace net {
+namespace tcp {
+
 ///////////////// CONCRETE STATES /////////////////
-using TCP = net::TCP;
-using Connection = TCP::Connection;
-using State = TCP::Connection::State;
+
+using State = Connection::State;
 
 /*
   CLOSED
@@ -49,7 +53,7 @@ public:
 
     => SynSent
   */
-  virtual Result handle(Connection&, TCP::Packet_ptr in) override;
+  virtual Result handle(Connection&, Packet_ptr in) override;
 
   inline virtual std::string to_string() const override {
     return "CLOSED";
@@ -79,7 +83,7 @@ public:
 
     => SynReceived.
   */
-  virtual Result handle(Connection&, TCP::Packet_ptr in) override;
+  virtual Result handle(Connection&, Packet_ptr in) override;
 
   inline virtual std::string to_string() const override {
     return "LISTENING";
@@ -108,7 +112,7 @@ public:
 
     => Established.
   */
-  virtual Result handle(Connection&, TCP::Packet_ptr in) override;
+  virtual Result handle(Connection&, Packet_ptr in) override;
 
   inline virtual std::string to_string() const override {
     return "SYN-SENT";
@@ -139,7 +143,7 @@ public:
 
     => Established.
   */
-  virtual Result handle(Connection&, TCP::Packet_ptr in) override;
+  virtual Result handle(Connection&, Packet_ptr in) override;
 
   inline virtual std::string to_string() const override {
     return "SYN-RCV";
@@ -167,7 +171,7 @@ public:
 
   virtual void abort(Connection&) override;
 
-  virtual Result handle(Connection&, TCP::Packet_ptr in) override;
+  virtual Result handle(Connection&, Packet_ptr in) override;
 
   inline virtual std::string to_string() const override {
     return "ESTABLISHED";
@@ -210,7 +214,7 @@ public:
 
     => FinWait2.
   */
-  virtual Result handle(Connection&, TCP::Packet_ptr in) override;
+  virtual Result handle(Connection&, Packet_ptr in) override;
 
   inline virtual std::string to_string() const override {
     return "FIN-WAIT-1";
@@ -246,7 +250,7 @@ public:
   /*
 
    */
-  virtual Result handle(Connection&, TCP::Packet_ptr in) override;
+  virtual Result handle(Connection&, Packet_ptr in) override;
 
   inline virtual std::string to_string() const override {
     return "FIN-WAIT-2";
@@ -288,7 +292,7 @@ public:
 
     => LastAck
   */
-  virtual Result handle(Connection&, TCP::Packet_ptr in) override;
+  virtual Result handle(Connection&, Packet_ptr in) override;
 
   inline virtual std::string to_string() const override {
     return "CLOSE-WAIT";
@@ -316,7 +320,7 @@ public:
 
     => TimeWait (Guess this isnt needed, just start a Close-timer)
   */
-  virtual Result handle(Connection&, TCP::Packet_ptr in) override;
+  virtual Result handle(Connection&, Packet_ptr in) override;
 
   inline virtual std::string to_string() const override {
     return "CLOSING";
@@ -346,7 +350,7 @@ public:
 
     => Closed (Tell TCP to remove this connection)
   */
-  virtual Result handle(Connection&, TCP::Packet_ptr in) override;
+  virtual Result handle(Connection&, Packet_ptr in) override;
 
   inline virtual std::string to_string() const override {
     return "LAST-ACK";
@@ -372,7 +376,7 @@ public:
   /*
 
    */
-  virtual Result handle(Connection&, TCP::Packet_ptr in) override;
+  virtual Result handle(Connection&, Packet_ptr in) override;
 
   inline virtual std::string to_string() const override {
     return "TIME-WAIT";
@@ -386,4 +390,7 @@ private:
   inline TimeWait() {};
 };
 
-#endif
+} // < namespace tcp
+} // < namespace net
+
+#endif // < NET_TCP_CONNECTION_STATES_HPP
