+++ api/net/tcp/rttm.hpp
@@ -1,0 +1,108 @@
+// License
+
+#pragma once
+#ifndef NET_TCP_RTTM_HPP
+#define NET_TCP_RTTM_HPP
+
+#include <os> // OS::uptime()
+#include <cassert>
+
+namespace net {
+namespace tcp {
+
+// [RFC 6298]
+// Round Trip Time Measurer
+struct RTTM {
+  using timestamp_t = double;
+  using duration_t = double;
+
+  // clock granularity
+  //static constexpr duration_t CLOCK_G = hw::PIT::frequency().count() / 1000;
+  static constexpr duration_t CLOCK_G = 0.0011;
+
+  static constexpr double K = 4.0;
+
+  static constexpr double alpha = 1.0/8;
+  static constexpr double beta = 1.0/4;
+
+  timestamp_t t; // tick when measure is started
+
+  duration_t SRTT; // smoothed round-trip time
+  duration_t RTTVAR; // round-trip time variation
+  duration_t RTO; // retransmission timeout
+
+  bool active = false;
+
+  RTTM()
+    : t(OS::uptime()), SRTT(1.0), RTTVAR(1.0), RTO(1.0), active(false)
+  {}
+
+  void start() {
+    t = OS::uptime();
+    active = true;
+  }
+
+  void stop(bool first = false) {
+    assert(active);
+    active = false;
+    // round trip time (RTT)
+    auto rtt = OS::uptime() - t;
+    debug2("<RTTM::stop> RTT: %ums\n",
+      (uint32_t)(rtt * 1000));
+    if(!first)
+      sub_rtt_measurement(rtt);
+    else {
+      first_rtt_measurement(rtt);
+    }
+  }
+
+  /*
+    When the first RTT measurement R is made, the host MUST set
+
+    SRTT <- R
+    RTTVAR <- R/2
+    RTO <- SRTT + max (G, K*RTTVAR)
+
+    where K = 4.
+  */
+  inline void first_rtt_measurement(duration_t R) {
+    SRTT = R;
+    RTTVAR = R/2;
+    update_rto();
+  }
+
+  /*
+    When a subsequent RTT measurement R' is made, a host MUST set
+
+    RTTVAR <- (1 - beta) * RTTVAR + beta * |SRTT - R'|
+    SRTT <- (1 - alpha) * SRTT + alpha * R'
+
+    The value of SRTT used in the update to RTTVAR is its value
+    before updating SRTT itself using the second assignment.  That
+    is, updating RTTVAR and SRTT MUST be computed in the above
+    order.
+
+    The above SHOULD be computed using alpha=1/8 and beta=1/4 (as
+    suggested in [JK88]).
+
+    After the computation, a host MUST update
+    RTO <- SRTT + max (G, K*RTTVAR)
+  */
+  inline void sub_rtt_measurement(duration_t R) {
+    RTTVAR = (1 - beta) * RTTVAR + beta * std::abs(SRTT-R);
+    SRTT = (1 - alpha) * SRTT + alpha * R;
+    update_rto();
+  }
+
+  inline void update_rto() {
+    RTO = std::max(SRTT + std::max(CLOCK_G, K * RTTVAR), 1.0);
+    debug2("<TCP::Connection::RTO> RTO updated: %ums\n",
+      (uint32_t)(RTO * 1000));
+  }
+
+}; // < struct RTTM
+
+} // < namespace tcp
+} // < namespace net
+
+#endif // < NET_TCP_RTTM_HPP
